# solution/backend/pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PROD Backend 2025 Advertising Platform API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/clients/bulk": {
            "post": {
                "description": "Создает или обновляет информацию о клиентах, переданных в запросе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Клиенты"
                ],
                "summary": "Создать или обновить клиента",
                "parameters": [
                    {
                        "description": "Список клиентов для создания или обновления",
                        "name": "clients",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Client"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Информация о клиентах, которые были успешно созданы или обновлены",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Client"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидные данные или повторяющиеся идентификаторы клиентов",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clients/{clientID}": {
            "get": {
                "description": "Возвращает информацию о клиенте по указанному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Клиенты"
                ],
                "summary": "Получить информацию о клиенте по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор клиента",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о клиенте",
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    },
                    "400": {
                        "description": "Неверный формат идентификатора клиента",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Клиент не найден",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/stats/advertisers/{advertiserID}/campaigns": {
            "get": {
                "description": "Возвращает агрегированную статистику по указанному рекламодателю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Статистика"
                ],
                "summary": "Получить статистику по рекламодателю",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор рекламодателя",
                        "name": "advertiserID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Агрегированная статистика по рекламодателю",
                        "schema": {
                            "$ref": "#/definitions/model.CampaignStats"
                        }
                    },
                    "400": {
                        "description": "Неверный идентификатор рекламодателя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Рекламодатель не найден",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats/advertisers/{advertiserID}/campaigns/daily": {
            "get": {
                "description": "Возвращает агрегированную статистику по рекламодателю за каждый день, начиная с даты создания рекламодателя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Статистика"
                ],
                "summary": "Получить дневную статистику по рекламодателю",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор рекламодателя",
                        "name": "advertiserID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Агрегированная статистика по рекламодателю за каждый день",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CampaignStatsDaily"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный идентификатор рекламодателя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Рекламодатель не найден",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats/campaigns/{campaignID}": {
            "get": {
                "description": "Возвращает агрегированную статистику по указанной кампании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Статистика"
                ],
                "summary": "Получить статистику по кампании",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор кампании",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Агрегированная статистика по кампании",
                        "schema": {
                            "$ref": "#/definitions/model.CampaignStats"
                        }
                    },
                    "400": {
                        "description": "Неверный идентификатор кампании",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Кампания не найдена",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats/campaigns/{campaignID}/daily": {
            "get": {
                "description": "Возвращает агрегированную статистику по кампаниям за каждый день от даты начала до текущего дня",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Статистика"
                ],
                "summary": "Получить дневную статистику по кампании",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор кампании",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Агрегированная статистика по кампании за каждый день",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CampaignStatsDaily"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Кампания не найдена",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CampaignStats": {
            "type": "object",
            "required": [
                "clicks_count",
                "conversion",
                "impressions_count",
                "spent_clicks",
                "spent_impressions",
                "spent_total"
            ],
            "properties": {
                "clicks_count": {
                    "type": "integer"
                },
                "conversion": {
                    "type": "number"
                },
                "impressions_count": {
                    "type": "integer"
                },
                "spent_clicks": {
                    "type": "number"
                },
                "spent_impressions": {
                    "type": "number"
                },
                "spent_total": {
                    "type": "number"
                }
            }
        },
        "model.CampaignStatsDaily": {
            "type": "object",
            "required": [
                "clicks_count",
                "conversion",
                "date",
                "impressions_count",
                "spent_clicks",
                "spent_impressions",
                "spent_total"
            ],
            "properties": {
                "clicks_count": {
                    "type": "integer"
                },
                "conversion": {
                    "type": "number"
                },
                "date": {
                    "type": "integer"
                },
                "impressions_count": {
                    "type": "integer"
                },
                "spent_clicks": {
                    "type": "number"
                },
                "spent_impressions": {
                    "type": "number"
                },
                "spent_total": {
                    "type": "number"
                }
            }
        },
...
        "model.Client": {
            "type": "object",
            "required": [
                "age",
                "client_id",
                "gender",
                "location",
                "login"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "client_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE"
                    ]
                },
                "location": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "required": [
                "message",
                "status"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}